:recycle: Move stock chart specific tests to own suite
